Adjust edit sort diff and behavior
The labels were reversed which made it confusing to read.

Also the trailing new line at the end of the file is a nice addition so
the file gets saved with a trailing new line. Without this, it would
leave your file in a state that's likely different than saving the file
with any code editor.