Add lint warning for incorrect amount sign
This isn't an error because it really depends on how you want to
organize your items. It could be correct in some cases but not others.

If you prefer putting both positive and negative amounts in 1 category,
like a business expense of -50.00 which then gets refunded for 50.00
then you end up with both positives and negatives in expenses.

But, this is still good to add as a warning to protect against legit
mistakes like putting -10.00 for an item you know you bought but didn't
return. You will have to use your discretion here to differentiate if you
need to address it.