Remove locale related tests
I don't know what's up with Github Actions yet. The locale is set there
to be the same as my machine which is ('en_US', 'UTF-8'). The setlocale
check isn't failing there so it's not falling back to $.

Yet on GitHub Actions the amounts are always displayed with $.

I will have to investigate this further but this feature appears to be
working outside of GitHub Actions. I'm able to set:

LC_ALL="el_GR.UTF8" plutus show

And it does show Greek formatted currencies.